feat: add rust wrapper for box.session.su, use C-API when available
Additionally:
- New `tarantool::session::user_id_by_name` API is available when running `picodata`.
- `tarantool::session::uid` and `tarantool::session::euid` when running `picodata` now use native C-API instead of calling into Lua.

There is a one breaking change: `tarantool::session::uid` and `tarantool::session::euid` now return `UserId` type which is an alias for `u32`. Previously `isize` was used.

+++ tarantool/src/schema/space.rs
@@ -34,7 +34,7 @@ pub fn create_space(name: &str, opts: &SpaceCreateOptions) -> Result<Space, Erro
 
     // Resolve ID of user, specified in options, or use ID of current session's user.
     let user_id = match &opts.user {
-        None => session::uid()? as u32,
+        None => session::uid()?,
         Some(user) => {
             let resolved_uid = schema::resolve_user_or_role(user.as_str())?;
             match resolved_uid {
